simplifies feature value assignation
